🔧 GL02 Final Fixes: Entity migration, Repository updates, Service improvements
✅ Key Changes:
- GL02Entity: Removed ITemporalEntity, proper NGAY_DL structure
- IGL02Repository: Entity-based interface with GL02Entity methods
- GL02Repository: Migrated to Repository<GL02Entity> base
- GL02Service: Complete Entity-first architecture
- DirectImportService: ParseGL02CsvAsync with GL02Entity mapping
- Repository fixes: long ID support, interface alignments

🏗️ Architecture:
- Entity-first throughout (no DataTable mixing)
- Partitioned Columnstore compliance
- Build: 0 errors, production ready

+++ Backend/TinhKhoanApp.Api/Controllers/LN03Controller.cs
@@ -133,7 +133,7 @@ namespace TinhKhoanApp.Api.Controllers
         public Task<bool> ExistsAsync(System.Linq.Expressions.Expression<Func<Models.DataTables.LN03, bool>> predicate) => Task.FromResult(_data.AsQueryable().Any(predicate));
         public Task<IEnumerable<Models.DataTables.LN03>> FindAsync(System.Linq.Expressions.Expression<Func<Models.DataTables.LN03, bool>> predicate) => Task.FromResult(_data.AsQueryable().Where(predicate).AsEnumerable());
         public Task<IEnumerable<Models.DataTables.LN03>> GetAllAsync() => Task.FromResult(_data.AsEnumerable());
-        public Task<Models.DataTables.LN03?> GetByIdAsync(int id) => Task.FromResult(_data.FirstOrDefault(x => x.Id == id));
+        public Task<Models.DataTables.LN03?> GetByIdAsync(long id) => Task.FromResult(_data.FirstOrDefault(x => x.Id == id));
         public Task<IEnumerable<Models.DataTables.LN03>> GetRecentAsync(int count) => Task.FromResult(_data.OrderByDescending(x => x.NGAY_DL).ThenByDescending(x => x.CREATED_DATE).Take(count).AsEnumerable());
         public void Remove(Models.DataTables.LN03 entity) { _data.Remove(entity); }
         public void RemoveRange(IEnumerable<Models.DataTables.LN03> entities) { foreach (var e in entities) _data.Remove(e); }
